easwars review 2 for xdsclient impl
